Added some empty view models for training schedule; added complete MonthViewModel

+++ MyCoach/MyCoach/DataHandling/DataTransferObjects/Month.cs
@@ -108,6 +108,6 @@ namespace MyCoach.DataHandling.DataTransferObjects
         /// <summary>
         ///     Ruft das Punkteziel für alle Monate ab, oder legt es fest.
         /// </summary>
-        public ushort TotalGoal { get; set; }
+        public int TotalGoal { get; set; }
     }
 }
